split the crate in 2 separate crates
This has been long overdue.

This split cosmogony into 2 crates:

- cosmogony_model with all the cosmogony types, that will be pushed to
crates.io
- cosmogony: only used by the cosmogony binary, used to create the
hierachy.

In addition to being cleaner, this remove the dependency on libgeos for
`cosmogony_model` :tada:
