+++ .gitignore
@@ -1,6 +1,6 @@
 # Generated by Cargo
 # will have compiled files and executables
-/target/
+target/
 
 # These are backup files generated by rustfmt
 **/*.rs.bk
