+++ cosmogony_model/src/lib.rs
@@ -1,0 +1,6 @@
+mod cosmogony;
+pub mod mutable_slice;
+mod zone;
+
+pub use crate::zone::{Coord, Zone, ZoneIndex, ZoneType};
+pub use cosmogony::{Cosmogony, CosmogonyMetadata, CosmogonyStats};
