+++ cosmogony_model/src/mutable_slice.rs
@@ -1,4 +1,4 @@
-use crate::zone::{Zone, ZoneIndex};
+use crate::{Zone, ZoneIndex};
 
 // This struct is necessary to wrap the `zones` slice
 // and keep a mutable reference to a zone (and set
