+++ src/bin/cosmogony.rs
@@ -1,6 +1,5 @@
-use cosmogony::cosmogony::Cosmogony;
 use cosmogony::{build_cosmogony, file_format::OutputFormat};
-use failure::Error;
+use cosmogony_model::Cosmogony;
 use flate2::write::GzEncoder;
 use flate2::Compression;
 use std::fs::File;
@@ -44,7 +43,10 @@ Accepted extensions are '.json', '.json.gz', '.jsonl', '.jsonl.gz'
     filter_langs: Vec<String>,
 }
 
-fn to_json_stream(mut writer: impl std::io::Write, cosmogony: &Cosmogony) -> Result<(), Error> {
+fn to_json_stream(
+    mut writer: impl std::io::Write,
+    cosmogony: &Cosmogony,
+) -> Result<(), failure::Error> {
     for z in &cosmogony.zones {
         serde_json::to_writer(&mut writer, z)?;
         writer.write(b"\n")?;
@@ -59,7 +61,7 @@ fn serialize_cosmogony(
     cosmogony: &Cosmogony,
     output_file: String,
     format: OutputFormat,
-) -> Result<(), Error> {
+) -> Result<(), failure::Error> {
     log::info!("writing the output file {}", output_file);
     let file = File::create(output_file)?;
     let stream = BufWriter::new(file);
@@ -82,7 +84,7 @@ fn serialize_cosmogony(
     Ok(())
 }
 
-fn cosmogony(args: Args) -> Result<(), Error> {
+fn cosmogony(args: Args) -> Result<(), failure::Error> {
     let format = OutputFormat::from_filename(&args.output)?;
 
     let cosmogony = build_cosmogony(
