+++ src/zone_typer.rs
@@ -1,4 +1,4 @@
-use crate::zone::{Zone, ZoneIndex, ZoneType};
+use cosmogony_model::{Zone, ZoneIndex, ZoneType};
 use failure::Fail;
 use failure::{err_msg, Error};
 use log::warn;
@@ -249,8 +249,8 @@ where
 #[cfg(test)]
 mod test {
     use super::CountryAdminTypeRules;
-    use crate::zone::{Zone, ZoneIndex, ZoneType};
     use crate::zone_typer::read_libpostal_yaml;
+    use cosmogony_model::{Zone, ZoneIndex, ZoneType};
 
     #[test]
     fn test_read_libpostal_yaml_basic() {
